requirements.txt
mss==9.0.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
pynput==1.7.6
pyqt5==5.15.10
cryptography==41.0.7
